Bugfixes to playlist handling
Ensure playlist shuffling works; I'm not seeing this behavior working as
expected with my current build. Not sure if that's a behavior change in
sqlite or if it just never worked; there's not much point in doing
the shuffling in the database, anyway, so just do it locally.

Support longer input lines when running acmd (1024 characters).

Looking at this code for the first time in years I think I was certainly
a little too excited to explore what could be offloaded to sqlite3; it
wouldn't surprise me if I started moving more code into the client.